1/8 22:16

+++ hours/lib/controller/owner_page_controller.dart
@@ -1,9 +1,11 @@
+import 'package:connectivity_plus/connectivity_plus.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:hours/core/function/split_data_to_tables.dart';
 import 'package:hours/core/function/time_format.dart';
 import 'package:hours/core/share/custom_snackbar.dart';
 
+import '../core/database/firebase_data.dart';
 import '../core/database/sqldb.dart';
 import '../core/function/calculate_time.dart';
 import '../core/model/employe_model.dart';
@@ -127,45 +129,83 @@ class OwnerPageControllerImp extends OwnerPageController {
 
   @override
   changeTimeValue(String column, String name, String value, int id) async {
+    //column:name of column in table Ex:Start At
+    //name:name of column in database Ex:startAt
+    //value: old value
+    //id:value of id in database
+
+    //Get new value from user
     TimeOfDay newTime = await showChangeTimeDialog(Get.context!, column, value);
     String newWorkTime = "00:00";
     String breakTime = "00:00";
+
+    //Check if there is a change in value
     if (value != timeFormat(newTime.hour, newTime.minute)) {
+      //update value in database
       await sqlDb.updateData(
           tableName, name, timeFormat(newTime.hour, newTime.minute), id);
+      //Get break time from database
       breakTime = await sqlDb.queryTime(tableName, "breakH", id);
+      //calculate new work time
       if (name == "startAt") {
+        //Get finish time from database
         String finishTime = await sqlDb.queryTime(tableName, "finishAt", id);
+        //calculate new time
+        //new work time=differance between(finish time,start time(new))-break time
         newWorkTime = subTime(
             differenceTime(
                 timeFormat(newTime.hour, newTime.minute), finishTime),
             breakTime);
       } else {
+        //Get Start time from database
         String startWork = await sqlDb.queryTime(tableName, "startAt", id);
+        //calculate new time
+        //new work time=differance between(finish time(new),start time)-break time
         newWorkTime = subTime(
             differenceTime(startWork, timeFormat(newTime.hour, newTime.minute)),
             breakTime);
       }
+      //save new work time
       await sqlDb.updateData(tableName, "workH", newWorkTime, id);
-
+      //update UI
       update();
+      //get  new record
+      Map newRow = await sqlDb.queryRow(tableName, id);
+      //upload new record
+      int upload = await uploadRecord(tableName, newRow);
+      //check if upload failed and update upload value in database to false
+      if (upload == 0) {
+        await sqlDb.updateData(tableName, "upload", "$upload", id);
+      }
     }
   }
 
   @override
   deleteRow(int id) {
+    //show confirm dialog to delete Row from table
     Get.defaultDialog(
         title: "Delete Row",
         titleStyle: const TextStyle(fontSize: 20),
         middleText: "Do you want delete the Row?",
         onCancel: () {},
         onConfirm: () async {
-          int respons = await sqlDb.deleteRow(tableName, id);
-          Get.back();
-          if (respons == 1) {
-            successfulSnackBar("the Row has been deleted");
+          final connectivityResult = await Connectivity().checkConnectivity();
+          if (connectivityResult != ConnectivityResult.none) {
+            //delete row from table in database
+            int respons = await sqlDb.deleteRow(tableName, id);
+            //delete row from firebase
+            await deleteRecord(tableName, id);
+            if (respons == 1) {
+              //show sucessful snackBar
+              Get.back();
+              successfulSnackBar("the Row has been deleted");
+              update();
+            }
+          } else {
+            await sqlDb.updateData(tableName, "upload", "2", id);
+            Get.back();
+            errorSnackBar("The row will be deleted when internet is available");
           }
-          update();
         });
   }
